+++ app/javascript/home/store/global/globalSlice.ts
