+++ app/javascript/home/store/sidebar/sidebarApi.ts
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export const fetchUsersApi = async () => {
+  return await axios.get('/api/users');
+};
